+++ internal/builtin/register_test.go
@@ -36,6 +36,7 @@ func TestRegister(t *testing.T) {
 		"osm:exec",
 		"osm:os",
 		"osm:path",
+		"osm:regexp",
 		"osm:time",
 		"osm:ctxutil",
 	}